+++ src/Utilities/Utility.h
@@ -22,13 +22,13 @@ char *joinStrings(const char *string1, size_t len1, const char *string2, size_t 
 
 char *joinStrings(const char *string1, const char *string2);
 
-double toSeconds(unsigned int gameTime);
+float toSeconds(unsigned int gameTime);
 
-unsigned int toMicroSeconds(double worldTime);
+unsigned int toMicroSeconds(float worldTime);
 
 unsigned int toMicroSeconds(unsigned int gameTime);
 
-unsigned int toGameTime(double worldTime);
+unsigned int toGameTime(float worldTime);
 
 unsigned int fromSDLTimeToGameTime(unsigned int SDLTime);
